Implement ConfigQueries trait and Postgres deployment ID support
Add write_config_snapshot to ConfigQueries trait and implement for ClickHouse,
Postgres, and DelegatingDatabaseConnection. Migrate all callsites from the free
function to the trait method. Add Postgres deployment_id table support with
ClickHouse fallback in howdy.
