+++ Cargo.toml
@@ -53,6 +53,7 @@ http = "1.4.0"
 infer = "0.19.0"
 tokio-stream = "0.1.15"
 tokio = { version = "1.48.0", features = ["full"] }
+tokio-util = { version = "0.7.17", features = ["rt"] }
 tracing = { version = "0.1.43", features = ["log"] }
 pyo3 = { version = "0.26.0", features = ["experimental-async", "abi3-py310"] }
 axum = { version = "0.8", features = ["macros", "http2"] }
@@ -115,7 +116,12 @@ durable-tools-spawn = { path = "internal/durable-tools-spawn" }
 tensorzero = { path = "clients/rust" }
 tensorzero-core = { path = "tensorzero-core" }
 regex = "1.12.2"
-redis = { version = "1.0.2", features = ["tokio-comp", "connection-manager", "tls-rustls", "tokio-rustls-comp"] }
+redis = { version = "1.0.2", features = [
+    "tokio-comp",
+    "connection-manager",
+    "tls-rustls",
+    "tokio-rustls-comp",
+] }
 redis-test = { version = "1.0.1", features = ["aio"] }
 globset = "0.4.18"
 walkdir = "2.5.0"
