+++ internal/durable-tools/src/tensorzero_client/embedded.rs
@@ -14,8 +14,8 @@ use tensorzero::{
     UpdateDatapointsResponse, WriteConfigRequest, WriteConfigResponse,
 };
 use tensorzero_core::config::snapshot::{ConfigSnapshot, SnapshotHash};
-use tensorzero_core::config::write_config_snapshot;
 use tensorzero_core::db::ConfigQueries;
+use tensorzero_core::db::delegating_connection::DelegatingDatabaseConnection;
 use tensorzero_core::db::feedback::FeedbackByVariant;
 use tensorzero_core::db::feedback::FeedbackQueries;
 use tensorzero_core::endpoints::datasets::v1::types::{
@@ -241,11 +241,13 @@ impl TensorZeroClient for EmbeddedClient {
 
         let hash = snapshot.hash.to_string();
 
-        write_config_snapshot(&self.app_state.clickhouse_connection_info, snapshot)
-            .await
-            .map_err(|e| {
-                TensorZeroClientError::TensorZero(TensorZeroError::Other { source: e.into() })
-            })?;
+        let db = DelegatingDatabaseConnection::new(
+            self.app_state.clickhouse_connection_info.clone(),
+            self.app_state.postgres_connection_info.clone(),
+        );
+        db.write_config_snapshot(&snapshot).await.map_err(|e| {
+            TensorZeroClientError::TensorZero(TensorZeroError::Other { source: e.into() })
+        })?;
 
         Ok(WriteConfigResponse { hash })
     }
