+++ tensorzero-core/src/db/clickhouse/mock_clickhouse_connection_info.rs
@@ -230,12 +230,17 @@ impl DatasetQueries for MockClickHouseConnectionInfo {
     }
 }
 
+#[async_trait]
 impl ConfigQueries for MockClickHouseConnectionInfo {
     async fn get_config_snapshot(
         &self,
         snapshot_hash: SnapshotHash,
     ) -> Result<ConfigSnapshot, Error> {
         self.config_queries.get_config_snapshot(snapshot_hash).await
+    }
+
+    async fn write_config_snapshot(&self, snapshot: &ConfigSnapshot) -> Result<(), Error> {
+        self.config_queries.write_config_snapshot(snapshot).await
     }
 }
