+++ tensorzero-core/src/db/clickhouse/mod.rs
@@ -10,11 +10,10 @@ use std::sync::Arc;
 use url::Url;
 
 use crate::config::BatchWritesConfig;
-use crate::config::snapshot::{ConfigSnapshot, SnapshotHash};
+use crate::db::HealthCheckable;
 use crate::db::clickhouse::clickhouse_client::ClickHouseClientType;
 use crate::db::clickhouse::clickhouse_client::DisabledClickHouseClient;
 use crate::db::clickhouse::clickhouse_client::ProductionClickHouseClient;
-use crate::db::{ConfigQueries, HealthCheckable};
 use crate::error::DelayedError;
 use crate::error::{Error, ErrorDetails};
 
@@ -30,6 +29,7 @@ use crate::db::clickhouse::clickhouse_client::FakeClickHouseClient;
 mod batch_inference;
 mod batching;
 pub mod clickhouse_client; // Public because tests will use clickhouse_client::FakeClickHouseClient and clickhouse_client::MockClickHouseClient
+pub mod config_queries;
 pub mod dataset_queries;
 pub mod dicl_queries;
 pub mod evaluation_queries;
@@ -331,46 +331,6 @@ impl Display for ClickHouseConnectionInfo {
 impl HealthCheckable for ClickHouseConnectionInfo {
     async fn health(&self) -> Result<(), Error> {
         self.inner.health().await
-    }
-}
-
-impl ConfigQueries for ClickHouseConnectionInfo {
-    async fn get_config_snapshot(
-        &self,
-        snapshot_hash: SnapshotHash,
-    ) -> Result<ConfigSnapshot, Error> {
-        #[derive(Deserialize)]
-        struct ConfigSnapshotRow {
-            config: String,
-            extra_templates: HashMap<String, String>,
-            #[serde(default)]
-            tags: HashMap<String, String>,
-        }
-
-        let hash_str = snapshot_hash.to_string();
-        let query = format!(
-            "SELECT config, extra_templates, tags \
-             FROM ConfigSnapshot FINAL \
-             WHERE hash = toUInt256('{hash_str}') \
-             LIMIT 1 \
-             FORMAT JSONEachRow"
-        );
-
-        let response = self.run_query_synchronous_no_params(query).await?;
-
-        if response.response.is_empty() {
-            return Err(Error::new(ErrorDetails::ConfigSnapshotNotFound {
-                snapshot_hash: hash_str,
-            }));
-        }
-
-        let row: ConfigSnapshotRow = serde_json::from_str(&response.response).map_err(|e| {
-            Error::new(ErrorDetails::ClickHouseDeserialization {
-                message: e.to_string(),
-            })
-        })?;
-
-        ConfigSnapshot::from_stored(&row.config, row.extra_templates, row.tags, &snapshot_hash)
     }
 }
