+++ tensorzero-core/src/db/mod.rs
@@ -171,12 +171,15 @@ pub trait ExperimentationQueries {
     ) -> Result<String, Error>;
 }
 
+#[async_trait]
 #[cfg_attr(test, automock)]
-pub trait ConfigQueries {
-    fn get_config_snapshot(
+pub trait ConfigQueries: Send + Sync {
+    async fn get_config_snapshot(
         &self,
         snapshot_hash: SnapshotHash,
-    ) -> impl Future<Output = Result<ConfigSnapshot, Error>> + Send;
+    ) -> Result<ConfigSnapshot, Error>;
+
+    async fn write_config_snapshot(&self, snapshot: &ConfigSnapshot) -> Result<(), Error>;
 }
 
 /// A stored DICL (Dynamic In-Context Learning) example.
