+++ tensorzero-core/src/db/postgres/mod.rs
@@ -10,7 +10,9 @@ use crate::error::{Error, ErrorDetails};
 use super::HealthCheckable;
 
 pub mod batch_inference;
+pub mod config_queries;
 pub mod dataset_queries;
+pub mod deployment_queries;
 pub mod experimentation;
 pub mod feedback;
 pub mod inference_queries;
